feat(core): Added package summary
Package summary includes count of internal, external and testing packages.
